requirements.txt
crewai
crewai-tools
litellm
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pandas
flask
python-dotenv
